fixed profile page url and navigate to this, fixed some test and ui screenshot tests, opened pre commit and ci
