+++ .github/workflows/main.yml
@@ -1,38 +1,38 @@
-#name: lint, test, build
-#run-name: ${{ github.actor }} is testing app
-#on:
-#  push:
-#    branches: [ main ]
-#  pull_request:
-#    branches: [ main ]
-#jobs:
-#  pipeline:
-#    runs-on: ubuntu-latest
-#    strategy:
-#      matrix:
-#        node-version: [ 20.x ]
-#    steps:
-#      - uses: actions/checkout@v2
-#      - name: Starting Node.js ${{ matrix.node-version }}
-#        uses: actions/setup-node@v1
-#        with:
-#          node-version: ${{ matrix.node-version }}
-#      - name: install modules
-#        run: npm install
-#      - name: build production project
-#        run: npm run build:prod
-#      - name: linting typescript
-#        run: npm run lint:ts
-#        if: always()
-#      - name: linting css
-#        run: npm run lint:scss
-#        if: always()
-#      - name: unit testing
-#        run: npm run test:unit
-#        if: always()
-#      - name: build storybook
-#        run: npm run storybook:build
-#        if: always()
-#      - name: screenshot testing
-#        run: npm run test:ui:ci
-#        if: always()
+name: lint, test, build
+run-name: ${{ github.actor }} is testing app
+on:
+  push:
+    branches: [ main ]
+  pull_request:
+    branches: [ main ]
+jobs:
+  pipeline:
+    runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        node-version: [ 20.x ]
+    steps:
+      - uses: actions/checkout@v2
+      - name: Starting Node.js ${{ matrix.node-version }}
+        uses: actions/setup-node@v1
+        with:
+          node-version: ${{ matrix.node-version }}
+      - name: install modules
+        run: npm install
+      - name: build production project
+        run: npm run build:prod
+      - name: linting typescript
+        run: npm run lint:ts
+        if: always()
+      - name: linting css
+        run: npm run lint:scss
+        if: always()
+      - name: unit testing
+        run: npm run test:unit
+        if: always()
+      - name: build storybook
+        run: npm run storybook:build
+        if: always()
+      - name: screenshot testing
+        run: npm run test:ui:ci
+        if: always()
