+++ .husky/pre-commit
@@ -1,9 +1,9 @@
 #!/usr/bin/env sh
 . "$(dirname -- "$0")/_/husky.sh"
 
-#npm run build:prod
-#npm run lint:ts
-#npm run lint:scss
-#npm run test:unit
-#npm run storybook:build
-#npm run test:ui:ci
+npm run build:prod
+npm run lint:ts
+npm run lint:scss
+npm run test:unit
+npm run storybook:build
+npm run test:ui:ci
