--- src/features/addCommentForm/ui/AddCommentForm/AddCommentForm.test.tsx
@@ -1,9 +1,0 @@
-import { AddCommentForm } from './AddCommentForm'
-import { render, screen } from '@testing-library/react'
-
-describe('AddCommentForm component', () => {
-    test('Create and test render', () => {
-        render(<AddCommentForm/>)
-        // expect(screen.getByText('TEST')).toBeInTheDocument()
-    })
-})
