+++ src/widgets/Sidebar/model/types/sidebar.ts
@@ -1,0 +1,9 @@
+import { type FC, type SVGProps } from 'react'
+
+export interface SidebarItemType {
+    id: number
+    path: string
+    text: string
+    Icon: FC<SVGProps<SVGSVGElement>>
+    authOnly?: boolean
+}
